+++ packages/frontend/src/components/legislator/legislator-list.tsx
@@ -1,5 +1,5 @@
 'use client'
-import React, { useMemo, useState, useCallback } from 'react'
+import React, { useMemo, useState, useCallback, useEffect } from 'react'
 import styled from 'styled-components'
 import useSWR from 'swr'
 import Link from 'next/link'
@@ -37,8 +37,6 @@ import FilterModal from '@/components/sidebar/filter-modal'
 import { fetchTopLegislatorsBySpeechCount } from '@/fetchers/legislator'
 // constants
 import { InternalRoutes } from '@/constants/navigation-link'
-// fetcher
-import fetchTopicOfALegislator from '@/fetchers/topic'
 // z-index
 import { ZIndex } from '@/styles/z-index'
 // lodash
@@ -114,6 +112,10 @@ const LegislatorList: React.FC<LegislatorListProps> = ({
     topics.map((topic) => topic) as TabProps[]
   )
 
+  useEffect(() => {
+    setTabList(topics.map((topic) => topic))
+  }, [topics])
+
   const selectedTopic = useMemo(() => {
     if (topics.length === 0) return null
     return topics[selectedTab] || topics[0]
@@ -181,7 +183,9 @@ const LegislatorList: React.FC<LegislatorListProps> = ({
       <Container>
         <Title $isEmpty={true} text="發言摘要" />
         <Body>
-          <Loader />
+          <EmptyState>
+            <Loader useAbsolute={false} />
+          </EmptyState>
         </Body>
       </Container>
     )
@@ -242,7 +246,6 @@ const LegislatorList: React.FC<LegislatorListProps> = ({
             title={`${legislatorName} 的相關發言篩選`}
             slug={legislatorSlug}
             initialSelectedOption={tabList}
-            fetcher={fetchTopicOfALegislator}
             onClose={() => {
               setShowFilter(false)
             }}