+++ packages/frontend/src/components/topic/topic-list.tsx
@@ -1,5 +1,5 @@
 'use client'
-import React, { useMemo, useState, useCallback } from 'react'
+import React, { useMemo, useState, useCallback, useEffect } from 'react'
 import styled from 'styled-components'
 import useSWR from 'swr'
 import Link from 'next/link'
@@ -113,6 +113,10 @@ const TopicList: React.FC<TopicListProps> = ({
     legislatorsData.map((legislator) => legislator) as TabProps[]
   )
 
+  useEffect(() => {
+    setTabList(legislatorsData.map((legislator) => legislator))
+  }, [legislatorsData])
+
   const selectedLegislator = useMemo(() => {
     if (legislatorsData.length === 0) return null
     return legislatorsData[selectedTab] || legislatorsData[0]
@@ -179,7 +183,9 @@ const TopicList: React.FC<TopicListProps> = ({
       <Container>
         <Title $isEmpty={true} text="發言摘要" />
         <Body>
-          <Loader />
+          <EmptyState>
+            <Loader useAbsolute={false} />
+          </EmptyState>
         </Body>
       </Container>
     )