+++ packages/frontend/src/fetchers/server/legislator.ts
@@ -12,6 +12,7 @@ export type LegislatorFromRes = {
     imageLink?: string
   }
   note?: string
+  tooltip?: string
   legislativeMeeting: {
     term: number
   }
@@ -150,6 +151,7 @@ export const fetchLegislator = async ({
           imageLink
         }
         note
+        tooltip
         legislativeMeeting {
           term
         }